+++ sentry-android-core/src/test/java/io/sentry/android/core/ManifestMetadataReaderTest.kt
@@ -813,4 +813,29 @@ class ManifestMetadataReaderTest {
         // Assert
         assertFalse(fixture.options.isAttachScreenshot)
     }
+
+    @Test
+    fun `applyMetadata reads send client reports to options`() {
+        // Arrange
+        val bundle = bundleOf(ManifestMetadataReader.CLIENT_REPORTS_ENABLE to false)
+        val context = fixture.getContext(metaData = bundle)
+
+        // Act
+        ManifestMetadataReader.applyMetadata(context, fixture.options)
+
+        // Assert
+        assertFalse(fixture.options.isSendClientReports)
+    }
+
+    @Test
+    fun `applyMetadata reads send client reports and keep default value if not found`() {
+        // Arrange
+        val context = fixture.getContext()
+
+        // Act
+        ManifestMetadataReader.applyMetadata(context, fixture.options)
+
+        // Assert
+        assertTrue(fixture.options.isSendClientReports)
+    }
 }