+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 
 Starting with version `6.6.0` of the `sentry`, [Sentry's version >= v21.9.0](https://github.com/getsentry/self-hosted/releases) is required or you have to manually disable sending client reports via the `sendClientReports` option. This only applies to self-hosted Sentry. If you are using [sentry.io](https://sentry.io), no action is needed.
 
+* Fix: Profiling rate decreased from 300hz to 100hz; fixed profiling traces folder creation on manual sdk init (#1997)
+* Fix: Allow disabling sending of client reports via Android Manifest and external options (#2007)
+
 ## 6.0.0-alpha.6
 
 * Feat: Add sentry-servlet-jakarta module (#1987)